Allow replacing reloads in `computeMinInitCons()`
"Reloads" meaning "reload states".

Add an overload of `computeMinInitCons()` where the set of reload states
of the input CMDP can be (conceptually) replaced by another set of
reload states. "Conceptually" because the input CMDP isn't actually
modified. Reason for adding this overload: in the algorithm for
computing Safe, MinInitCons needs to be computed for a CMDP with a
modified set of reload states.

Also put the parameters of some functions on lines of their own because
the signatures are very long.