Set Safe(r) to zero for good reload states r
Suppose the agent starts in a reload state r. If survival can be
guaranteed with some initial resource level d <= capacity, then the
minimum initial resource level that is required, is in fact zero, since
the initial state is a reload state and so the resource level is set to
the capacity after leaving the initial state.