Define helper function for max over successors
Define a helper function to take the maximum of a vector of resource
levels, where we only consider the vector's values for states which are
a potential successor of a given state when taking a given action. I
will need to take such a maximum in another function besides
`computeMinInitCons()`, so putting it in a separate helper function
reduces code duplication.