Change `maxOverSuccessors()` parameter order
I will define a function later which corresponds to the (mathematical)
function "SPR-Val" from the paper I'm using. SPR-Val has as parameters a
state, an action, and a vector of resource levels. I want to be
consistent with that order and I want all my functions to be consistent.
Hence this change.