Read target states from PRISM file
Output of algorithm for SafePR seems correct now. (The vector SafePR
anyway. Still haven't looked at the counter selector.)

The problem was most likely that when manually setting the target states
in the C++ code, I used the index of the target state as it is in the
original CMDP. However, Storm changed the order of the states, so the
index I used didn't actually correspond to the state that was supposed
to be the target state.