Add TeeStream and duplicate output to a file
I did not create `TeeStream`. I copied it from
https://wordaligned.org/articles/cpp-streambufs
and changed the formatting a bit, as well as the names and namespace.

When running the programme, some standard storm-related output is
printed. This is inconvenient when you want to compare the output with
the output of another programme, especially for a large number of
inputs. By constructing a `TeeStream` that outputs to `std::cout` and a
file, and using that `TeeStream` only for my own output, I get my own
output in that file, without the unwanted output.

This commit also changes `std::cout` in
`storm::cmdp::printCounterSelector()` to the `ostream` reference
parameter `out`, which is a mistake I had made earlier. I didn't notice
because I passed `std::cout` as `out` anyway.