Add function to encode resource level into states
I have not yet tried compiling this code.
    `storm::cmdp::getTransitionMatrixAccordingToCounterSelector()`
returns a transition matrix where the states correspond to
pairs (state, resourceLevel). There is also a special state which
represents the agent being out of energy. Transitions correspond to what
the provided counter selector would choose and the transitions take care
of the cost and recharging.