Add function to get target labels for new MDP
Consider transforming a CMDP into an MDP where the states represent
pairs (s, rl) where s is a state from the original CMDP and rl is a
resource level. Each pair (s, rl) should be labelled as a target state
if s is a target state. This is done by
`storm::cmdp::getStateLabellingForBuiltInResourceLevels()`.