Clean up
Due to time constraints I had to evaluate the performance of my
implementation before being able to finish these TODOs. I have already
presented a plot with runtimes of my implementation, so at this point I
would rather not change the implementation in a way that might affect
the runtime performance.