Leave state order
Initially, I was applying a function to the result vectors to sort them
based on the PRISM variable determining the state (and I assumed that
there was only one such variable). This was useful to make the output
easier to understand. But later I started using PRISM models where the
state depends on multiple variables, which meant that the ordering
resulting from sorting the vectors didn't make much more sense than the
original ordering. What's more, by leaving the state order, the order is
the same for my implementation and the implementation [1] against which
I wanted to compare the results. This way, comparing them is much
easier. The reason for the orderings being the same for both
implementations is most likely that they both use Storm.

[1] https://github.com/FiMDP/FiMDP